Refactoring the Directory fidget by extracting components, utilities, and configuration into separate modules
