+++ src/fidgets/token/Directory/utils.ts
@@ -149,3 +149,12 @@ export function sortMembers(
   return entries;
 }
 
+// Re-export utilities from sub-modules
+export { parseCsv, chunkArray } from "./utils/csv";
+export {
+  getNestedUser,
+  mapNeynarUserToMember,
+  createDefaultMemberForCsv,
+  type NeynarUser,
+} from "./utils/memberData";
+
